refactor: rename cmp for more clarity (#1119)
