+++ src/frontend/src/icp/components/tokens/IcAddTokenReview.svelte
